Move model registration to side-specific event handler
